after 3-nd code review:  - refactored 'init_db.sql': set DECIMAL in column 'price'

+++ src/main/resources/init_db.sql
@@ -3,6 +3,6 @@ USE book;
 CREATE TABLE `books` (
 `id` BIGINT AUTO_INCREMENT,
 `title` VARCHAR(255),
-`price` BIGINT,
+`price` DECIMAL,
 PRIMARY KEY(`id`)
 );
